# Collection, Lst<T>, LnkLst interface cleanup / refactor