There were some issues with the current interface hierarchy surrounding collections:

- `LnkLst` derived publicly from `Lst<ObjKey>`, but uses a different "index space", because `LnkLst` hides tombstone links (unresolved links), while `Lst<ObjKey>` doesn't. This meant that it was easy to accidentally call a method on `LnkLst` and get or use indices that don't make sense or are out of bounds. `LnkLst::find_first` and `LnkLst::find_all` had this problem.
- `Lst<T>` could not be marked `final` due to the inheritance, which prevents devirtualization (inlining, direct method call, etc.).
- Several methods were statically overridden, rather than dynamically, resulting in dispatch based on the type of pointer you had, rather than the actual type of the instance. This could lead to similar bugs.
- Members on `CollectionBase` and friends made it difficult / ugly to compose a collection implementation from other collections (say, implementing `LnkLst` in terms of `Lst<ObjKey>` without deriving).
- Handling of tombstones / unresolved links was isolated to `LnkLst`, and could not be reused in other collections.
- `Lst<T>` had a weird and poorly justified precondition that all calls must be preceded by a call to `size()`, in order to get the accessor properly refreshed. Failure to do so was only detected in debug mode, and could result in corruption. The motivation behind it was to avoid a bounds check in some methods, but at high risk.

This refactor makes the following changes:

- All member variables move to the bottom-most class in the class hierarchy.
- Methods on base classes are marked as `virtual`.
- ... but concrete collection types are marked as `final`, so stuff can be inlined.
- Several interfaces have been cleaned up, such that more things are private and fewer things can leak through layers.
- All methods on `Lst<T>` and other collections call `update_if_needed()`. Will measure performance impact; none expected.
- The `bump_content_version()` / `update_if_needed()` flow is updated such that a collection accessor never has a precondition of calling some methods before others just to refresh the accessor.
    - Some appropriate inlining is added to ensure that the overhead is minimal in the common case.
- It is impossible to get a `Lst<ObjKey>&` that is actually a `LnkLst` (avoiding the mix of dynamic and static overrides). In other words, it is impossible to access the contents of a `LnkLst` without accounting for tombstones.
     - Use `Obj::get_linklist()` to get a `LnkLst` (condensed indices).
     - Use `Obj::get_list<ObjKey>()` get a `Lst<ObjKey>` (uncondensed indices).
     - `Obj::get_listbase_ptr()` returns a `LnkLst` instance.
- Stable iteration over `Lst<T>` is removed (the `m_deleted` member of `CollectionBase`). This is unused by SDKs and added complexity.
- `Collection<T, Interface>` is gone, its responsibilities divided between `CollectionImplBase<Interface>` and concrete collection implementations (`Lst<T>`, `Set<T>`, `Dictionary`).
- `CollectionBase` no longer derives from `ArrayParent`.
- `LnkLst` no longer derives from `Lst<ObjKey>`, removing the risk of mixing condensed/uncondensed indices. The choice between access by condensed or uncondensed indices is made at instantiation time, rather than at method dispatch time.
- `real2virtual()` is added as a companion to `virtual2real()`, which is needed for find() functions. Opposed to `virtual2real()`, it is implemented using binary search.
- Collection iterators are now `random_access_iterator`s, rather than `bidirectional_iterator`s. This is important for utility functions such as `std::lower_bound`.

The new class hierarchy looks like this for `LnkLst` (from the bottom up):

- `LnkLst` (final, internally has a `Lst<ObjKey>` as a member)
    - `ObjCollectionBase<LstBase>` (tombstone handling, `virtual2real` etc.)
        - `LstBase` (base class for all lists)
           - `CollectionBase` (base class for all collections)
        - `ObjList` (generic interface for lists of objects)

The new class hierarchy looks like this for `Lst<T>` (from the bottom up):

- `Lst<T>` (final, internally has `BPlusTree<T>`)
    - `CollectionBaseImpl<LstBase>` (convenience implementation of things shared by all containers (`m_obj`, `m_col_key`, etc.)).
        - `ArrayParent` (protected base class, used to attach the collection accessor to the object)
        - `LstBase` (base class for all lists)
            - `CollectionBase` (base class for all collections)

The new class hierarchy for `Dictionary` (from the bottom up):

- `Dictionary`
    - `CollectionBaseImpl<CollectionBase>`
        - `ArrayParent`
        - `CollectionBase`

The new class hierarchy for `Set<T>` (from the bottom up):

- `Set<T>`
    - `CollectionBaseImpl<SetBase>`
        - `ArrayParent`
        - `SetBase`
            - `CollectionBase`

Goals achieved:

- Protection against nonsensical list indices due to tombstone handling.
- Centralized implementation of tombstone handling.
- Classes are `final`, enabling devirtualization.
- Implementation details are hidden where possible.
- The `ArrayParent` interface is out of sight.