@simonask 2020-11-11
# Performance

Possible effects:

- Since all `Lst<T>` methods now implicitly call `update_if_needed()`, there is an additional atomic load for each of those calls, and the potential call to `init_from_parent()`. Note that this call would happen anyway with the old implementation, but it would be front-loaded to the call to `size()`. However, users that maintain the old usage pattern will have a call to both `size()` followed by `insert()`/`set()`/`remove()`/etc., so there is a penalty of one extra atomic load.
- On the other hand, since `Lst<T>` is marked `final`, no virtual method calls should happen when going through the concrete type (i.e., not through a base class interface, such as `LstBase` or `CollectionBase`).

## Benchmark

The benchmark is based on `Lst<int64_t>`, and run on an AMD Ryzen 3950X with GCC 9 on Ubuntu 20.04 under WSL2 on Windows 10. The build is a standard `Release` build (no LTO).

### Before

```
1000000 repetitions of insert
    time: 144 ns/rep
1000000 repetitions of interlaced get/add
    time: 542 ns/rep
```

### After

```
1000000 repetitions of insert
    time: 113 ns/rep
1000000 repetitions of interlaced get/add
    time: 492 ns/rep
```

## Analysis

The new version is slightly faster on this hardware, but it's likely that the penalties of virtual method calls versus atomic loads have a different ratio on other hardware (especially ARM hardware). However, the difference is unlikely to be in orders of magnitude.