@simonask 2020-11-13
Reviewers: I added a bunch of documentation for `CollectionBase`, hoping to set a precedent. ;-)